core(scoring): extract from reportgenerator into separate file
