+++ lighthouse-core/report/v2/report-generator.js
@@ -44,23 +44,6 @@ class ReportGeneratorV2 {
     return REPORT_TEMPLATES;
   }
 
-  /**
-   * Computes the weighted-average of the score of the list of items.
-   * @param {!Array<{score: number|undefined, weight: number|undefined}>} items
-   * @return {number}
-   */
-  static arithmeticMean(items) {
-    const results = items.reduce((result, item) => {
-      const score = Number(item.score) || 0;
-      const weight = Number(item.weight) || 0;
-      return {
-        weight: result.weight + weight,
-        sum: result.sum + score * weight,
-      };
-    }, {weight: 0, sum: 0});
-
-    return (results.sum / results.weight) || 0;
-  }
 
   /**
    * Replaces all the specified strings in source without serial replacements.
@@ -81,35 +64,6 @@ class ReportGeneratorV2 {
         .join(firstReplacement.replacement);
   }
 
-  /**
-   * Returns the report JSON object with computed scores.
-   * @param {{categories: !Object<string, {id: string|undefined, weight: number|undefined, audits: !Array<{id: string, weight: number|undefined}>}>}} config
-   * @param {!Object<{score: ?number|boolean|undefined}>} resultsByAuditId
-   * @return {{score: number, categories: !Array<{audits: !Array<{score: number, result: !Object}>}>}}
-   */
-  generateReportJson(config, resultsByAuditId) {
-    const categories = Object.keys(config.categories).map(categoryId => {
-      const category = config.categories[categoryId];
-      category.id = categoryId;
-
-      const audits = category.audits.map(audit => {
-        const result = resultsByAuditId[audit.id];
-        // Cast to number to catch `null` and undefined when audits error
-        let auditScore = Number(result.score) || 0;
-        if (typeof result.score === 'boolean') {
-          auditScore = result.score ? 100 : 0;
-        }
-
-        return Object.assign({}, audit, {result, score: auditScore});
-      });
-
-      const categoryScore = ReportGeneratorV2.arithmeticMean(audits);
-      return Object.assign({}, category, {audits, score: categoryScore});
-    });
-
-    const overallScore = ReportGeneratorV2.arithmeticMean(categories);
-    return {score: overallScore, categories};
-  }
 
   /**
    * Returns the report HTML as a string with the report JSON and renderer JS inlined.
